+++ HCDProject/Assets/Scripts/NPC/Monster/MonsterState/MonsterAttackState.cs
@@ -1,0 +1,38 @@
+using UnityEngine;
+
+public class MonsterAttackState : IState
+{
+    private BaseMonster _controller;
+    private float _timer;
+
+    public MonsterAttackState(BaseMonster controller)
+    {
+        _controller = controller;
+    }
+    
+    public void Enter()
+    {
+        _timer = 0f;
+    }
+
+    public void Update()
+    {
+        if (Vector2.Distance(_controller.transform.position, _controller.GetTargetObject.transform.position) >=
+            _controller.Stats._attackRange || _controller.GetTargetObject == null)
+        {
+            _controller.CurrentState.Value = EStateType.Chase;
+            return;
+        }
+        
+        _timer += Time.deltaTime;
+        if (_timer >= 2f)
+        {
+            _controller.UseSkill((int)ESkillType.Normal);
+        }
+    }
+    
+    public void Exit()
+    {
+
+    }
+}
