+++ HCDProject/Assets/Scripts/NPC/Monster/MonsterState/MonsterIdleState.cs
@@ -11,7 +11,7 @@ public class MonsterIdleState : IState
     
     public void Enter()
     {
-        
+        _controller.SetCurrentTarget(_controller.Detect(_controller.Stats._chaseRange));
     }
     
     public void Update()
