--- HCDProject/Assets/Scripts/NPC/Monster/MonsterState/MonsterNearAttackState.cs
@@ -1,26 +1,0 @@
-using UnityEngine;
-
-public class MonsterNearAttackState : IState
-{
-    private BaseMonster _controller;
-
-    public MonsterNearAttackState(BaseMonster controller)
-    {
-        _controller = controller;
-    }
-    
-    public void Enter()
-    {
-
-    }
-
-    public void Update()
-    {
-        
-    }
-    
-    public void Exit()
-    {
-
-    }
-}
